fix: ensure that CLI help is formatted deterministically and set defaults are displayed properly
