+++ snakemake/common/argparse.py
@@ -1,4 +1,5 @@
 import argparse
+import dataclasses
 
 import configargparse
 
@@ -55,3 +56,32 @@ def register_parser_action(parse_func, kwargs):
             setattr(namespace, self.dest, parsed)
 
     kwargs["action"] = ParserAction
+
+
+class ArgumentDefaultsHelpFormatter(argparse.HelpFormatter):
+    """Help message formatter which adds default values to argument help.
+
+    Like argparse.ArgumentDefaultsHelpFormatter, but doesn't print
+    None/dataclasses._MISSING_TYPE/etc.
+    """
+
+    def _get_help_string(self, action):
+        if (
+            (
+                action.option_strings
+                or action.nargs in [argparse.OPTIONAL, argparse.ZERO_OR_MORE]
+            )
+            and action.default not in (None, "", set(), argparse.SUPPRESS)
+            and not isinstance(action.default, dataclasses._MISSING_TYPE)
+        ):
+            if isinstance(action.default, frozenset):
+                # ensure deterministic sorting and proper display without braces
+                # and commas
+                return (
+                    action.help
+                    + f" (default: {' '.join(sorted(map(str, action.default)))})"
+                )
+            else:
+                return action.help + " (default: %(default)s)"
+        else:
+            return action.help
